backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
slack-bolt>=1.20.0
slack-sdk>=3.30.0
aiohttp>=3.9.0
anthropic>=0.39.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
